requirements.txt
google-genai>=1.0.0
google-adk>=0.3.0
nemoguardrails>=0.10.0
openai>=1.0.0
litellm>=0.5.0
detoxify>=0.5.0
python-dotenv>=1.0.0